+++ PlayBuilder/Program.cs
@@ -16,6 +16,8 @@ builder.Services.AddSingleton<ICollectionRuleService, AtlasCollectionRuleService
 builder.Services.AddSingleton<IAtlasComparisonService, AtlasComparisonService>();
 builder.Services.AddSingleton<ICollectionService, CollectionService>();
 builder.Services.AddSingleton<IGameScanWorkflowService, GameScanWorkflowService>();
+builder.Services.AddSingleton<IGameScanSessionService, GameScanSessionService>();
+builder.Services.AddSingleton<IFolderPickerService, WindowsFolderPickerService>();
 
 // Atlas is registered as a parallel decision engine during its staged migration.
 builder.Services.AddSingleton<FilenameTokenizer>();